pyproject.toml
[tool.pdm]

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.3.0",
]
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "", email = ""},
]
dependencies = [
    "djangorestframework>=3.14.0",
    "django-cors-headers>=3.14.0",
    "python-magic>=0.4.27",
    "Pillow>=9.5.0",
    "django-environ>=0.10.0",
    "mysqlclient>=2.1.1",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"
